% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boreal_data.R
\docType{data}
\name{boreal}
\alias{boreal}
\title{A sample food-web network from the boreal region of the Barents Sea}
\format{A data frame with 2 columns and 1546 rows. Each row is a node pair
   where the first element of each node pair is 'consumed' (or pollinated
   etc) by the second element.}
\source{
Kortsch, S. et al. 2015. Climate change alters the structure of
   arctic marine food webs due to poleward shifts of boreal generalists.
   Proc. R. Soc. B 282: 20151546.
}
\usage{
data(boreal)
}
\description{
The sample network is a marine food web from the boreal region of the Barents
Sea. The Barents Sea is a large, open sub-arctic shelf sea bordering the
Arctic Ocean. The sample food web contains 180 trophospecies, consisting of
detritus, 8 basal taxa, 32 zooplankton, 66 benthic, 56 fish, 8 sea birds and
9 marine mammal, and 1546 feeding interactions.
}
\examples{
data(boreal)
}
\keyword{datasets}
