% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_group_element.R
\name{close_group_element}
\alias{close_group_element}
\title{Given a group and an element not in the group, close the group}
\usage{
close_group_element(group, new_element)
}
\arguments{
\item{group}{sign matrix with group elements in columns}

\item{new_element}{n-vector with signs that is not in the group}
}
\value{
the closed group
}
\description{
Given a group and an element not in the group, close the group
}
\examples{
close_group_element(matrix(c(1, 1, 1, -1)), c(-1, -1))

}
\keyword{close}
\keyword{element}
\keyword{group}
