% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_group_generators.R
\name{close_group_generators}
\alias{close_group_generators}
\title{Construct group from generators}
\usage{
close_group_generators(generators)
}
\arguments{
\item{generators}{sign matrix with generators in columns}
}
\value{
the closed group
}
\description{
Construct group from generators
}
\examples{
close_group_generators(construct_oracle_generators(8))

}
\keyword{combine}
\keyword{generators}
