% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_0_generators.R
\name{construct_0_generators}
\alias{construct_0_generators}
\title{Construct the generators for a 0-subgroup}
\usage{
construct_0_generators(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
matrix with generators for a 0-subgroup as rows
}
\description{
Construct the generators for a 0-subgroup
}
\examples{
construct_0_generators(4)
}
\keyword{0-subgroup}
\keyword{generators}
