% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_0_subgroup.R
\name{construct_0_subgroup}
\alias{construct_0_subgroup}
\title{Construct a 0-subgroup}
\usage{
construct_0_subgroup(n)
}
\arguments{
\item{n}{number of observations that is a power of 2}
}
\value{
a 0-subgroup as an n x n matrix, with elements in columns
}
\description{
Constructs a single 0-subgroup for n observations
}
\examples{
construct_0_subgroup(4)
}
\keyword{0-subgroup}
\keyword{construct}
