% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_candidate_elements.R
\name{construct_candidate_elements}
\alias{construct_candidate_elements}
\title{Construct potential elements that could be added to the group}
\usage{
construct_candidate_elements(group, num_candidates)
}
\arguments{
\item{group}{the group}

\item{num_candidates}{number of candidates generated for each group expansion}
}
\value{
a set of candidate elements
}
\description{
Construct potential elements that could be added to the group
}
\examples{
construct_candidate_elements(construct_oracle_subgroup(4), 10)

}
\keyword{candidate}
\keyword{elements}
