% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_near_oracle_subgroup.R
\name{construct_near_oracle_subgroup}
\alias{construct_near_oracle_subgroup}
\title{Construct a near oracle subgroup}
\usage{
construct_near_oracle_subgroup(
  n,
  num_candidates,
  max_order,
  leak_fun,
  group = NULL
)
}
\arguments{
\item{n}{number of signs per sign-flip element}

\item{num_candidates}{number of candidates generated with each group expansion}

\item{max_order}{maximum order of the group}

\item{leak_fun}{the function used to compare groups}

\item{group}{# optional. A starting group. Default is just the identity group}
}
\value{
a near oracle group
}
\description{
Construct a near oracle subgroup
}
\examples{
construct_near_oracle_subgroup(4, 100, 8, leak_max)

}
\keyword{group}
\keyword{near}
\keyword{oracle}
