% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_neg_subgroup.R
\name{construct_neg_subgroup}
\alias{construct_neg_subgroup}
\title{Construct a negative subgroup}
\usage{
construct_neg_subgroup(group)
}
\arguments{
\item{group}{a sign-flipping group}
}
\value{
its negative subgroup
}
\description{
Construct a negative subgroup
}
\examples{
construct_neg_subgroup(construct_oracle_subgroup(4))
}
\keyword{construct}
\keyword{negative}
\keyword{subgroup}
