% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_oracle_generators.R
\name{construct_oracle_generators}
\alias{construct_oracle_generators}
\title{Construct the generators for an oracle subgroup}
\usage{
construct_oracle_generators(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
matrix with generators for an oracle subgroup as rows. Identity in column 1
}
\description{
Construct the generators for an oracle subgroup
}
\examples{
construct_oracle_generators(4)
}
\keyword{generators}
\keyword{oracle}
\keyword{subgroup}
