% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_oracle_subgroup.R
\name{construct_oracle_subgroup}
\alias{construct_oracle_subgroup}
\title{Construct an oracle subgroup}
\usage{
construct_oracle_subgroup(n)
}
\arguments{
\item{n}{number of observations that is a power of 2}
}
\value{
an oracle subgroup as an n x n matrix, with elements in columns
}
\description{
Constructs a single oracle subgroup for n observations
}
\examples{
construct_oracle_subgroup(4)
}
\keyword{construct}
\keyword{oracle}
\keyword{subgroup}
