% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_sign_flips.R
\name{duplicated_sign_flips}
\alias{duplicated_sign_flips}
\title{Duplicated sign-flips}
\usage{
duplicated_sign_flips(sign_matrix)
}
\arguments{
\item{sign_matrix}{a matrix of sign-flips}
}
\value{
duplicate elements
}
\description{
For a given matrix with sign-flip columns, this returns the indices
of the duplicate columns
}
\examples{
duplicated_sign_flips(matrix(c(1, 1, 1, -1, 1, -1), ncol = 3))

}
\keyword{duplicate}
\keyword{sign-flip}
