% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_2_edge.R
\name{freqMat_2_edge}
\alias{freqMat_2_edge}
\title{Convert a frequency interaction matrix to an edge list}
\usage{
freqMat_2_edge(x)
}
\arguments{
\item{x}{A frequency interaction matrix, in which rows are species that are
consumed (or pollinated etc) and columns are those species that are the
consumers (or pollinators etc). For example, for a plant-pollinator
interaction matrix, the rows would be the plant species, and the columns
the pollinator species. The interaction matrix can be a dataframe or a
matrix. Each value should be numeric, indicating the number of interactions
between two species. The names of the species must be set as the rownames
and column names of the interaction matrix prior to running the function.
As the functions in the 'nos' R package are based on presence-absence data,
all interactions greater than or equal to 1 are set to 1. Cannibalistic
interactions are allowed, i.e. the same species can be in a row and column.}
}
\value{
An edge list in the form of a dataframe. Each value in a column is a
  node (e.g. a food item in a trophic-web). Nodes are set as numeric values.
  For each row (i.e. node pair), the value in the first column is 'consumed'
  (or pollinated, parasitized etc) by the value in the second column.
}
\description{
Converts frequency interaction matrices, as used with the
  \link[bipartite]{bipartite} R package, into edge lists. As the functions in the 'nos' R
  package are based on presence-absence data, all interactions greater than
  or equal to 1 are set to 1.
}
\examples{
sim_dat <- matrix(c(0, 2, 1, 0, 1, 2, 0, 1), ncol = 4) #simulate food web matrix
rownames(sim_dat) <- c("A", "B") #name the consumed species
colnames(sim_dat) <- c("C", "D", "A", "E") #name the consumer species (with one cannibal interaction)
freqMat_2_edge(sim_dat)
}
