% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sign_flipping_group.R
\name{generate_sign_flipping_group}
\alias{generate_sign_flipping_group}
\title{Generate the sign-flipping group}
\usage{
generate_sign_flipping_group(n)
}
\arguments{
\item{n}{number of observations}
}
\value{
the sign-flipping group in dimension n
}
\description{
Generate the sign-flipping group
}
\examples{
generate_sign_flipping_group(8)

}
\keyword{group}
\keyword{sign-flipping}
