% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak_l1.R
\name{leak_l1}
\alias{leak_l1}
\title{Leak l1}
\usage{
leak_l1(group)
}
\arguments{
\item{group}{a matrix of sign-flips}
}
\value{
the l1 leak value
}
\description{
Leak l1
}
\examples{
leak_l1(construct_oracle_subgroup(8))

}
\keyword{l1}
\keyword{leak}
