% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak_linf.R
\name{leak_linf}
\alias{leak_linf}
\title{Leak l-infinity}
\usage{
leak_linf(leak_terms)
}
\arguments{
\item{group}{a matrix of sign-flips}
}
\value{
the l-infinity leak value
}
\description{
Leak l-infinity
}
\examples{
leak_linf(colSums(construct_oracle_subgroup(8)))

}
\keyword{abs}
\keyword{leak}
\keyword{linf}
\keyword{max}
