% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak_mean.R
\name{leak_mean}
\alias{leak_mean}
\title{Leak mean}
\usage{
leak_mean(leak_terms)
}
\arguments{
\item{leak_terms}{a vector of a sum of sign-flips}
}
\value{
the mean leak value
}
\description{
Leak mean
}
\examples{
leak_mean(colSums(construct_oracle_subgroup(8)))

}
\keyword{leak}
\keyword{mean}
