% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak_minimizing_group_expansion.R
\name{leak_minimizing_group_expansion}
\alias{leak_minimizing_group_expansion}
\title{Construct group expansion that minimizes the leak}
\usage{
leak_minimizing_group_expansion(group, num_elements, leak_fun)
}
\arguments{
\item{group}{the group}

\item{num_elements}{number of candidates generated for each group expansion}

\item{leak_fun}{the function used to compare groups}
}
\value{
an expanded group
}
\description{
Construct group expansion that minimizes the leak
}
\examples{
leak_minimizing_group_expansion(construct_oracle_subgroup(4), 10, leak_max)

}
\keyword{expand}
\keyword{group}
\keyword{leak}
