% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak_minimizing_group_expansion_batched.R
\name{leak_minimizing_group_expansion_batched}
\alias{leak_minimizing_group_expansion_batched}
\title{Construct group expansion that minimizes the leak using batches to reduce
memory usage}
\usage{
leak_minimizing_group_expansion_batched(
  group,
  num_elements,
  leak_fun,
  batch_size = min(1000, num_elements)
)
}
\arguments{
\item{group}{the group}

\item{num_elements}{number of candidates generated for each group expansion}

\item{leak_fun}{the function used to compare groups}
}
\value{
an expanded group
}
\description{
Construct group expansion that minimizes the leak using batches to reduce
memory usage
}
\examples{
leak_minimizing_group_expansion_batched(construct_oracle_subgroup(4), 10, leak_max, 2)


}
\keyword{batched}
\keyword{expand}
\keyword{group}
\keyword{leak}
