% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{NOSM_dir}
\alias{NOSM_dir}
\title{Compute N_bar using a directed network without a user provided network
of potential interactions}
\usage{
NOSM_dir(net, perc = 1, sl = 1)
}
\arguments{
\item{net}{A network, in the form of an edge list. This should be a matrix or
dataframe with two columns. Each value in a column is a node (e.g. a food
item in a trophic-web). Nodes can be identified using numbers or
characters. For each row (i.e. node pair), the value in the first column is
'consumed' (or pollinated, parasitized etc) by the value in the second
column.}

\item{perc}{(default to 1) - the fraction of node pair comparisons to be
performed to compute N_bar. We recommend performing all possible pair
comparisons (perc = 1). However, for exploratory analyses on large sets of
networks (or for very large networks), the possibility of using a lower
fraction of pair comparisons is a useful option.}

\item{sl}{(default is 1) Specifies whether cannibalistic interactions should
be considered as possible and therefore taken into account and removed
during computation ('1') or not ('0').}
}
\value{
A list (two elements) of class 'NOSM' with a 'Type' attribute 'Dir'.
  The first element in the list is a vector of overlap values for the "in
  nodes" and the second element is a vector of overlap values for the "out
  nodes".

  The \code{\link{summary.NOSM}} methods provides more useful summary
  statistics.
}
\description{
Computation of N_bar (NOS) using a directed network (e.g. food
  web) and without a user provided network of potential interactions. In a
  directed, unimode network, all nodes having at least an in-coming link will
  be considered as potential partners for nodes having at least an out-going
  link and vice-versa.
}
\examples{
data(boreal)
x <- NOSM_dir(boreal, perc = 1, sl = 1)
summary(x)
}
