% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unwanted_elements.R
\name{remove_unwanted_elements}
\alias{remove_unwanted_elements}
\title{Removes unwanted elements from the set of sign-flips}
\usage{
remove_unwanted_elements(elements, group)
}
\arguments{
\item{elements}{a matrix of sign-flips as columns}

\item{group}{a group which the elements should not be in}
}
\value{
a set of elements with the unwanted elements removed
}
\description{
Removes unwanted elements from the set of sign-flips
}
\keyword{elements}
\keyword{remove}
\keyword{sign-flip}
\keyword{unwanted}
