% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testList_data.R
\docType{data}
\name{testList}
\alias{testList}
\title{Unit tests data}
\format{
An object of class \code{list} of length 4.
}
\usage{
testList
}
\description{
A dataset used in the unit tests. Has four elements, the first three
being input data, and the fourth the output data. All generated using
the python nos package to ensure the results match up.
}
\examples{
data(testList)
}
\keyword{datasets}
