% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boreal_data.R
\docType{data}
\name{boreal}
\alias{boreal}
\title{A sample *** network from **boreal**}
\format{A data frame with 2 columns and 1546 rows. Each row is a node pair
   where the first element of each node pair is 'consumed' (or pollinated
   etc) by the second element.}
\source{
Kortsch, S. et al. 2015. Climate change alters the structure of
   arctic marine food webs due to poleward shifts of boreal generalists.
   Proc. R. Soc. B 282: 20151546.
}
\usage{
data(boreal)
}
\description{
A **Boreal food-web network of ** species in ***.
}
\examples{
data(boreal)
}
\keyword{datasets}
