% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_near_oracle_subgroup_path.R
\name{construct_near_oracle_subgroup_path}
\alias{construct_near_oracle_subgroup_path}
\title{Construct a path of near oracle subgroups}
\usage{
construct_near_oracle_subgroup_path(n, num_candidates, max_rank, leak_fun)
}
\arguments{
\item{n}{number of signs per sign-flip element}

\item{num_candidates}{number of candidates generated with each group expansion}

\item{leak_fun}{the function used to compare groups}

\item{max_order}{maximum rank (log2(order)) of the group}

\item{group}{# optional. A starting group. Default is just the identity group}
}
\value{
a near oracle group
}
\description{
Construct a path of near oracle subgroups
}
\examples{
construct_near_oracle_subgroup_path(4, 100, 10, leak_max)

}
\keyword{group}
\keyword{near}
\keyword{oracle}
