% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_oracle_subgroup.R
\name{construct_oracle_subgroup}
\alias{construct_oracle_subgroup}
\title{Constructs the largest possible oracle subgroup}
\usage{
construct_oracle_subgroup(n, max_order = n)
}
\arguments{
\item{n}{number of observations that is a power of 2}

\item{max_order}{maximum order of the oracle subgroup. Default is n}
}
\value{
an oracle subgroup as an p x n matrix, with elements in columns
}
\description{
Constructs an oracle subgroup for n observations of desired order,
given that it exists
}
\examples{
construct_oracle_subgroup(4)
}
\keyword{construct}
\keyword{oracle}
\keyword{subgroup}
