% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leak_max.R
\name{leak_max}
\alias{leak_max}
\title{Leak max}
\usage{
leak_max(group)
}
\arguments{
\item{group}{a matrix of sign-flips}
}
\value{
the max leak value
}
\description{
Leak max
}
\examples{
leak_max(construct_oracle_subgroup(8))

}
\keyword{leak}
\keyword{max}
